requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LangChain and AI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-text-splitters>=1.0.0
openai>=1.10.0

# Vector database
chromadb>=0.4.22

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Optional: Anthropic (commented out by default)
# anthropic>=0.8.0
